Struct cpupower::Cpu

pub struct Cpu {
    // some fields omitted
}

Methods

impl Cpu

fn get_all() -> Iterator

Iterate over all Cpu's available in your system

fn exists(id: CpuId) -> bool

Check whether a Cpu with given ID exists in you system

fn new(id: CpuId) -> Cpu

fn get_freq(&self) -> Result<Frequency>

Get frequency reported by hardware or by kernel This function tries to get freq using call to hardware first, and if that call fails - uses call to kernel

fn get_freq_kernel(&self) -> Result<Frequency>

Get frequency reported by your kernel According to the underlying library documentation - you don't need to be root to perform this operation

fn get_freq_hardware(&self) -> Result<Frequency>

Get frequency reported by your hardware According to the underlying library documentation - you should be root to perform this operation

fn set_freq(&self, freq: Frequency) -> Result<&Cpu>

Set frequency for the given CPU You should have root privileges to do that

fn get_transition_latency(&self) -> Result<u64>

fn modify_policy_max(&self, max: Frequency) -> Result<()>

fn modify_policy_min(&self, min: Frequency) -> Result<()>

fn modify_policy_governor(&self, governor: &str) -> Result<()>

fn get_hardware_limits(&self) -> Result<(Frequency, Frequency)>

fn get_id(&self) -> CpuId

Get if of current processor

fn get_driver(&self) -> String

fn get_policy(&self) -> Result<Policy>

fn set_policy(&self, policy: &Policy) -> Result<()>

fn get_available_governors(&self) -> Result<Vec<String>>

fn get_available_frequencies(&self) -> Result<Vec<Frequency>>

fn get_affected_cpus(&self) -> Result<Vec<Cpu>>

fn get_stats(&self) -> Result<Vec<Stat>>